refactor(discovery): decompose the snmp collector god-file by role (WS-D)
Split internal/discovery/snmp_collector.go (688 LOC) into role-based files
in the same package — pure move: types + collector orchestration/dispatch
stay in snmp_collector.go; the per-area low-level collect* implementations
move out.

All files under the 600-line size-gate threshold. Declaration sets and
code-line multisets verified identical to the original.

Related to #187
